establish connection extend db by model
